cronopete.vala: Exclude special folders
There are several folders that shouldn't be backed up. Examples are:

~/.cache: it contains temporary cached files (like the web browser cache)
~/.dbus: contains the DBus sockets, so its contents aren't backup-able
~/.gvfs: (in old systems) contains FUSE-mounted remote file systems
~/.var/app/*/cache: contains cached files from Flatpack applications

Until now, marking the option to backup the hidden files in the HOME
folder meant that, unless the user manually added those folders to the
list of folders to skip.

This patch changes this, and ensures that those folders won't be backed
up.