backup_rsync.vala: Fix bug when not enough disk for single backup
If there isn't enough disk space for a single backup (this is, if the
first backup is so big that it doesn't fit in the destination disk),
cronopete will retry up to the infinity and beyond, instead of stop
and show an error message.

This patch fixes this by ensuring that, when trying to delete an old
backup to get more free space, there must remain at least one full
backup; if not, an error message will be shown to the user, informing
(s)he that the disk used for backups is too small.

Ensuring that there remains at least one full backup also improves
the security of cronopete, because it would be very risky if storing
a new backup requires to delete all the previous backups in the
disk: if, after deleting all of them to obtain enough free space,
a problem happens with the backup, there won't remain any backup,
and another failure in the main disk will mean loosing all the data.