Better management of rsync error codes
This patch implements a more complete management of the exit codes of
RSYNC. As before, receiving an 11 exit code will be managed as
"no free disk space". But now the 24 exit code (some files vanished)
will be ignored, because there is no reason to bother the user in
this case. Also, the exit code 23 is managed. Now, if there are less
that 10 megabytes of free disk space, it will be presumed that there
was a problem with it, and an old backup will be erased before
retrying the operation. But if there are more than 10 megabytes, the
error will be shown to the user.